requirements.txt
requests==2.32.3
tenacity==8.5.0
beautifulsoup4==4.12.3
pg8000==1.31.2
python-dateutil==2.9.0.post0
pydantic==2.8.2
pytest==8.2.2
pytest-mock==3.12.0
requests-toolbelt==1.0.0